+++ app/assets/stylesheets/module-dashboards-maker.scss
