+++ app/assets/stylesheets/module-dashboards-viewer.scss
