+++ app/javascript/gobierto_dashboards/index.js
