+++ app/javascript/gobierto_dashboards/modules/dashboard_commons_controller.js
