+++ app/javascript/gobierto_dashboards/modules/dashboard_viewer_controller.js
