+++ app/javascript/gobierto_dashboards/webapp/Maker.vue
@@ -62,7 +62,7 @@
 
 <script>
 // add the styles here, because this element can be inserted both as a component or standalone
-import "../../../assets/stylesheets/module-TEMP-maker.scss";
+import "../../../assets/stylesheets/module-dashboards-maker.scss";
 import Viewer from "./Viewer";
 import HeaderForm from "./components/HeaderForm";
 import Aside from "./layouts/Aside";
