+++ app/javascript/gobierto_dashboards/webapp/components/SmallCard.vue
