+++ app/javascript/gobierto_dashboards/webapp/components/Widget.vue
