+++ app/javascript/gobierto_dashboards/webapp/components/WidgetEditable.vue
