+++ app/javascript/gobierto_dashboards/webapp/components/WidgetHTML.vue
