+++ app/javascript/gobierto_dashboards/webapp/components/WidgetIndicator.vue
