+++ app/javascript/gobierto_dashboards/webapp/components/WidgetIndicatorIndividual.vue
