+++ app/javascript/gobierto_dashboards/webapp/components/WidgetIndicatorTable.vue
