+++ app/javascript/gobierto_dashboards/webapp/lib/factories.js
