+++ app/javascript/gobierto_dashboards/webapp/lib/mock__DELETABLE.js
