+++ app/javascript/gobierto_dashboards/webapp/lib/widgets.js
