+++ app/javascript/packs/admin.js
@@ -9,5 +9,5 @@
 
 import 'lib/commons'
 import 'gobierto_admin'
-import 'gobierto_TEMP'
+import 'gobierto_dashboards'
 import 'lib/i18n'
