+++ app/views/gobierto_admin/gobierto_dashboards/dashboards/show.html.erb
