+++ app/views/sandbox/admin_dashboards.erb
@@ -1,0 +1,15 @@
+<% content_for :javascript_module_link do %>
+  <!-- No es necesario adjuntarlo xq ya está DENTRO del pack 'admin' -->
+  <!-- <%= javascript_packs_with_chunks_tag 'dashboards' %> -->
+<% end %>
+
+<% content_for :stylesheet_module_link do %>
+  <%= stylesheet_packs_with_chunks_tag 'dashboards' %>
+<% end %>
+
+<a data-dashboards-maker class="button open_remote_modal" href="/admin/gobierto_dashboards/index">Nuevo</a>
+<a data-dashboards-maker class="button open_remote_modal" href="/admin/gobierto_dashboards/show">Editar</a>
+
+<hr>
+
+<div id="dashboard-viewer-app" data-context="CONTEXTO"></div>
