+++ app/views/sandbox/dashboards.erb
@@ -1,9 +1,9 @@
 <% content_for :javascript_module_link do %>
-  <%= javascript_packs_with_chunks_tag 'TEMP' %>
+  <%= javascript_packs_with_chunks_tag 'dashboards' %>
 <% end %>
 
 <% content_for :stylesheet_module_link do %>
-  <%= stylesheet_packs_with_chunks_tag 'TEMP' %>
+  <%= stylesheet_packs_with_chunks_tag 'dashboards' %>
 <% end %>
 
 <div class="column">
